chore: fix ci actions (#2)

+++ src/Component/Http/Runtime/Middleware/AccessLogMiddleware.php
@@ -2,6 +2,15 @@
 
 declare(strict_types=1);
 
+/*
+ * This file is part of the Neutomic package.
+ *
+ * (c) Saif Eddin Gmati <azjezz@protonmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Neu\Component\Http\Runtime\Middleware;
 
 use Neu\Component\Http\Message\RequestInterface;
@@ -31,7 +40,7 @@ final readonly class AccessLogMiddleware implements PrioritizedMiddlewareInterfa
     {
         $time = microtime(true);
         $response = $next->handle($context, $request);
-        $duration = microtime(true) - $time;
+        $duration = (int) (microtime(true) - $time);
 
         $message = Str\format(
             '"%s %s" %d HTTP/%s @ %s | %s ms',
